requirements.txt
# Core LangChain packages
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.24
langchain-pinecone==0.2.11

# Embeddings / Transformers (PyTorch only)
sentence-transformers==3.3.1
transformers[torch]==4.45.2
huggingface-hub==0.26.2
torch>=2.3.0

# Vector Databases
pinecone==7.3.0
#  memory 
langgraph

# Frameworks / APIs
flask==3.1.1

# Utils
pypdf==5.6.1
python-dotenv==1.1.0

# Groq SDK
groq==0.9.0
langchain-groq==0.1.5

# Editable local install
-e .
